Fix troglobit#131: Regression in last hop router compared to v2.3.2.
The new v3 pimd uses integer arrays for tracking oifs rather than a bit array in v2.3.2.  The bit array oifs and leaves comparison is a simple C language AND function, which quickly checks for the presence of at least one matching oif or leave.  The new arrays have a little more work to loop through the array comparison.